Session one
With a bunch of components, and api routes connected to them. The
overall feel of the app has been set up, and it is integrated with
a Redux store.

Tweaked packages in both subfolders to accommodate the needs that
have emerged.

Style structure established, started mobile-first with all the UI.